Moved SelectSignalJets in SignalObjectSelector and related methods. Fixed some compilation errors
