+++ Analysis/include/SignalObjectSelector.h
@@ -9,6 +9,7 @@ This file is part of https://github.com/hh-italian-group/h-tautau. */
 #include "h-tautau/Cuts/include/hh_bbtautau_2017.h"
 #include "h-tautau/Cuts/include/H_tautau_2016_baseline.h"
 #include "h-tautau/Cuts/include/H_tautau_2017_baseline.h"
+#include "h-tautau/JetTools/include/BTagger.h"
 
 namespace analysis {
 
@@ -25,12 +26,103 @@ ENUM_NAMES(SignalMode) = {
     { SignalMode::TauPOG_Skimmer, "TauPOG_Skimmer" }
 };
 
+namespace jet_ordering {
+
+    template<typename LorentzVector>
+    struct JetInfo {
+        LorentzVector p4;
+        size_t index;
+        double tag;
+
+        JetInfo() : index(0), tag(0.0) { }
+
+        JetInfo(const LorentzVector _p4, size_t _index, double _tag)
+            : p4(_p4), index(_index), tag(_tag) { }
+
+    };
+
+    template<typename LorentzVector>
+    bool CompareJets(const JetInfo<LorentzVector>& jet_1, const JetInfo<LorentzVector>& jet_2,
+                     double pt_thr, double eta_thr)
+    {
+        const auto eta1 = std::abs(jet_1.p4.eta());
+        const auto eta2 = std::abs(jet_2.p4.eta());
+        if(eta1 < eta_thr && eta2 >= eta_thr) return true;
+        if(eta1 >= eta_thr && eta2 < eta_thr) return false;
+        const auto pt1 = jet_1.p4.pt();
+        const auto pt2 = jet_2.p4.pt();
+        if(pt1 > pt_thr && pt2 <= pt_thr) return true;
+        if(pt1 <= pt_thr && pt2 > pt_thr) return false;
+
+        if(jet_1.tag != jet_2.tag)
+            return jet_1.tag > jet_2.tag;
+        return pt1 > pt2;
+    };
+
+    template<typename LorentzVector>
+    std::vector<JetInfo<LorentzVector>> OrderJets(
+                                  const std::vector<JetInfo<LorentzVector>>& jet_info_vector,
+                                  bool apply_hard_cut,
+                                  double pt_cut = std::numeric_limits<double>::lowest(),
+                                  double eta_cut = std::numeric_limits<double>::max())
+    {
+        const auto comparitor = [&](const JetInfo<LorentzVector>& jet_1,
+                                    const JetInfo<LorentzVector>& jet_2) -> bool {
+            return analysis::jet_ordering::CompareJets(jet_1, jet_2, pt_cut, eta_cut);
+        };
+
+        std::vector<JetInfo<LorentzVector>> jets_ordered;
+        if(apply_hard_cut){
+            for(size_t n = 0; n < jet_info_vector.size(); ++n) {
+                if(jet_info_vector.at(n).p4.Pt() > pt_cut && std::abs(jet_info_vector.at(n).p4.eta()) < eta_cut)
+                    jets_ordered.push_back(jet_info_vector.at(n));
+            }
+        }
+        else
+            jets_ordered = jet_info_vector;
+        std::sort(jets_ordered.begin(), jets_ordered.end(), comparitor);
+        return jets_ordered;
+    }
+
+}
+
 class SignalObjectSelector {
 public:
+    using JetPair = ntuple::JetPair;
+
     SignalObjectSelector(SignalMode _mode);
 
     bool PassLeptonSelection(const ntuple::TupleLepton& lepton, Channel channel) const;
     boost::optional<size_t> GetHiggsCandidateIndex(const ntuple::Event& event) const;
+
+    struct SelectedSignalJets{
+        JetPair selectedBjetPair;
+        JetPair selectedVBFjetPair;
+        size_t n_bjets;
+
+        SelectedSignalJets();
+        bool HasBjetPair(size_t njets) const;
+        bool HasVBFPair(size_t njets) const;
+        bool isSelectedBjet(size_t n) const;
+        bool isSelectedVBFjet(size_t n) const;
+    };
+
+    static SelectedSignalJets SelectSignalJets(const ntuple::Event& event,
+                                               const analysis::Period& period,
+                                               analysis::JetOrdering jet_ordering,
+                                               size_t selected_higgs_index);
+
+     template<typename LorentzVector>
+     static bool PassEcalNoiceVetoJets(const LorentzVector& jet_p4, Period period, int jets_pu_id)
+     {
+         if(period !=  analysis::Period::Run2017)
+             return true;
+
+         const double abs_eta = std::abs(jet_p4.eta());
+         return !(jet_p4.pt() < cuts::hh_bbtautau_2017::jetID::max_pt_veto &&
+                     abs_eta > cuts::hh_bbtautau_2017::jetID::eta_low_veto &&
+                     abs_eta < cuts::hh_bbtautau_2017::jetID::eta_high_veto && (jets_pu_id & (1 << 2)) == 0);
+     }
 
 private:
     bool PassHTT_LeptonSelection(const ntuple::TupleLepton& lepton, Channel channel, bool is_sync) const;
@@ -38,6 +130,7 @@ private:
     bool PassHH_LeptonSelection(const ntuple::TupleLepton& lepton, Channel channel) const;
     bool PassSkimmer_LeptonSelection(const ntuple::TupleLepton& lepton) const;
     bool PassTauPOG_Skimmer_LeptonSelection(const ntuple::TupleLepton& lepton) const;
+
 
 
 private:
